// src/navigation/types.ts
export type MyToursStackParamList = {
  MyToursList: undefined;
  TourDetail: { tourId: string };
};

export type MapStackParamList = {
  MapView: undefined;
  TourDetail: { tourId: string };
};

export type RootTabParamList = {
  Home: undefined;
  MyTours: undefined;
  Map: undefined;
  Contact: undefined;
  Profile: undefined;
};